戰鬥協議實裝

+++ Assets/Scripts/GameModule/Wrapper.cs
@@ -27,7 +27,7 @@ public class Wrapper
         return UnityEngine.Random.Range(min, max);
     }
 
-    static public float GetRandom(int min, int max)
+    static public int GetRandom(int min, int max)
     {
         return UnityEngine.Random.Range(min, max);
     }
